backend/requirements.txt
flask
flask-cors
nltk
openai
python-dotenv